+++ yellowbox/__init__.py
@@ -1,4 +1,6 @@
 from yellowbox._version import __version__
+from yellowbox.clients import docker_client
+from yellowbox.image_build import image_build
 from yellowbox.networks import connect, temp_network
 from yellowbox.retry import RetrySpec
 from yellowbox.service import YellowService
@@ -6,6 +8,8 @@ from yellowbox.subclasses import ContainerService, RunMixin, SingleContainerServ
 
 __all__ = [
     '__version__',
+    'docker_client',
+    'image_build',
     'connect', 'temp_network',
     'RetrySpec',
     'YellowService',